+++ test/td_cache/taxonomy_cache_test.exs
@@ -1,8 +1,12 @@
 defmodule TdCache.TaxonomyCacheTest do
   use ExUnit.Case
+
   alias TdCache.Redix
   alias TdCache.Redix.Stream
+  alias TdCache.AclCache
   alias TdCache.TaxonomyCache
+
+  @role "test_role"
 
   doctest TdCache.TaxonomyCache
 
@@ -141,6 +145,31 @@ defmodule TdCache.TaxonomyCacheTest do
     assert %{id: ^id1, parent_ids: [^id1], external_id: _, name: _} = map[id1]
     assert %{id: ^id2, parent_ids: [^id2, ^id1], external_id: _, name: _} = map[id2]
     assert %{id: ^id3, parent_ids: [^id3, ^id2, ^id1], external_id: _, name: _} = map[id3]
+  end
+
+  describe "has_role?/4" do
+    setup %{
+      root: root,
+      parent: %{id: id2} = parent,
+      domain: domain
+    } do
+      user_id = System.unique_integer([:positive])
+      Enum.each([root, parent, domain], &TaxonomyCache.put_domain/1)
+      AclCache.set_acl_role_users("domain", id2, @role, [user_id])
+      on_exit(fn -> AclCache.delete_acl_roles("domain", id2) end)
+      [user_id: user_id]
+    end
+
+    test "returns true if a user_id has a role in a domain or its parents", %{
+      root: %{id: id1},
+      parent: %{id: id2},
+      domain: %{id: id3},
+      user_id: user_id
+    } do
+      refute TaxonomyCache.has_role?(id1, @role, user_id)
+      assert TaxonomyCache.has_role?(id2, @role, user_id)
+      assert TaxonomyCache.has_role?(id3, @role, user_id)
+    end
   end
 
   defp random_domain(params \\ %{}) do